ENT-4220: Added fix for threaded_queue_test hanging
Test was sometimes hanging on rhel7 because of
test_threads_wait_empty. When the waiting thread was spawned,
ThreadedQueueWaitEmpty occasionally managed to acquire the lock
before any of the threads that were spawned to add elements, which
resulted in the amount of elements added were 101, not 100 like
the while loop was checking for.

Another sleep statement was added to let the pushing threads
finish their work before spawning the waiting thread.

Ticket: ENT-4220
Changelog: None