Add colorization to cf-diag.sh
Add colorization to cf-diag to make it easier for operators to find and fix problems.

It only colorizes the output if stdout is pointing to a tty.

We could have gotten fancier and added an option to force color regardless, but I don't think that's necessary.

This change was made by Mike Weilgart