Fix dcs_check_regcmp when not expecting difference in comparison
This bundle from the dcs test framework has not been used with a negative check.
It appears that this dcs_check_regcmp bundle is a derivative of
dcs_check_strcmp, but it was never edited to replace equal with matched and
thus, has never worked properly in the case where you do not expect a mismatch
and don't find one.