ENT-3832: Added thread safe double-edged queue (deque)
Can push and pop left/right, pop multiple elements in one go,
report count, capacity and if it is empty. It has pretty much
the same innards as ThreadedQueue, and will expand if capacity
is reached. Every operation is thread-safe (apart from deleting,
which requires locks to be unused), and it supports waiting for
the deque to be empty, as well as for it to be filled with
elements.

Ticket: ENT-3832
Changelog: None