Added a new utility to contrib: cf-remote
cf-remote is a python + fabric tool to log in to remote hosts
you have ssh access to. It will be used to download, transfer,
and install cfengine packages as well as bootstrapping etc.
This first iteration includes the `info` command, which prints
basic information about a remote host.

Ticket: CFE-2889
Changelog: Commit
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>