Quote the package module path when passing it to the interpreter
If the interpreter for a package module is specified, the package
module itself is passed as an argument to the interpreter (a
script path). If it contains spaces (or some other special
characters) it needs to be quoted.

Ticket: ENT-4432
Changelog: None
(cherry picked from commit cb711f15fb4b0b7bd7ebd160477488d0c9347950)