ENT-3832: Added tests for threaded_deque
Tests pushing and popping to and from both left and right,
pushing and popping NULL and empty deques, copying and
pushing/popping to the copied deque, checking count, expanding
and operating on expanded deques, popping multiple elements in
one operation, and concurrent operations from multiple threads.
Also tests the thread conditions while waiting for the deque to
empty, as well as it not being empty.

Ticket: ENT-3832
Changelog: None