CFE-2678: Global connection cache
Currently the connection cache is reset after each bundle pass.
This limits its effectivity, as all policies do not group file copies
in the same bundle pass. This was apprently done to limit the risk
of reusing broken connections.

This commit keeps a unique connection cache for the whole agent run,
and adds an error detection mechanism to avoid reusing broken cached
connections.