Modified StringCopy function to indicate when result was truncated
The function now takes buffer size instead of string length as an
argument. It still returns string length, except in case
truncation happened, it will return buf_size (strlen + 1).

This allows the caller to tell whether the string fit the buffer
perfectly or if it was truncated. Also rearranged the arguments
so they make sense (from, to, buf_size).

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>