Don't allow NULL pointers as strings in hashing
NULL pointer is not a string and should not be accepted
as keys or for hashing. A NULL usually indicates that
something went wrong (a function call failed). If the
error is not handled (by checking for NULL) it will be
used in subsequent function calls, like HashString().
To prevent this and force programmers to handle errors
correctly, don't accept NULL as a string.

See corresponding commit in enterprise.

Changelog: None
Ticket: ENT-4505
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit 7483de23ea157a7c9092f99cdf132176d2067500)