Private keys generated by cf-key are no longer encrypted
Private key files encrypted with a broken cipher and default
hard coded passphrase provide no real security, and is only an
inconvenience. Maybe it was intended to add a password prompt
later, but it's been 10 years now, and the cipher and passphrase
remain untouched. The function which reads keys still supports
both encrypted and unencrypted keys, it will decrypt if necessary.

This passphrase was introduced as part of cf-key in:
c7765f71ff04f07f497988a2a068122394da30bc

Changelog: Commit
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>