Make sure map->load is 0 at the end of HashMapClear()
Otherwise the size is unchanged and the map thinks it still has
all the items that were there before. The items are gone, but
when new ones are inserted, the table sooner or later hits the
threshold and grows automatically. Repeat this process of "insert
a couple hundred items, clear, repeat" multiple times and the
table grows and grows...and grows.

Let's decrement map->load everytime we remove an item and assert
that it reaches 0 when all items are removed.

Ticket: CFE-3032
Changelog: Fix growing memory footprint of daemons