Added --no-truncate option to cf-key
This option, when used with --show-hosts changes the formatting
of the output. Instead of padding and truncating each of the
fields, they are printed, in full, with no padding, and separated
by a single tab character. The output is not as pretty, but should
be more useful for parsing by other scripts / tooling.

Changelog: Commit
Ticket: CFE-3036
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>