Unresolved function calls in process_select body are now skipped
Function calls which always fail, like getuid("nosuchuser"), are
never resolved. Previously this would cause a programming error,
since the body is expected to have a list of strings, not
unresolved function calls.

The function calls are silently skipped (with a verbose message)
as this matches the behavior of calling the functions in a vars
promise, and using that as a body parameter.

Changelog: Commit
Ticket: CFE-1968
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit 942ec326921acca37a598862cb123a1d22bdb6db)