Add 'tname' to 'ps' output fields on Linux
So that it can be used for filtering processes.

Portability note:

"libpromises/systype.c", which specifies the OS-specific
arguments to the invoked "ps" command, had only been requesting
the "TTY" header for a small number of platforms: sun and
solaris.

As an aside, note that "ps -eo ...tname..." and "ps -eo
...tty..."  are subtly different.

Remaining OSes will also have this new failure, for which the
solution ought to be an equivalent change in "systype.c".

Ticket: CFE-3025
Changelog: Title
Edited-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>