LGTM: Detect non-bool functions which return true or false
This custom query looks for functions with return type other than
`int` and `bool`, and alerts if they return `true` or `false`.

The severity can be reduced later, I just want to test this.
It can also be expanded to catch int functions, but then
we will have a lot of alerts.

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>