fix bogus success message for 'cfe_internal' package
Commit 9cd859620825 ("Fix cfe_internal packages to avoid return code
check.") had ExecPackageCommand() skip the return code verification to
avoid noise in the log for a command that is expected to fail.  This
results in ExecPackageCommand() indicating success, though, which leads
to the completely bogus message:

  verbose: Package schedule execution ok for 'cfe_internal_non_existing_package' (outcome cannot be promised by cf-agent)

Decide if we want to ignore the package before looking at the status
code to restore the intended behavior.