Allow 1 process with TIME not between 0s and 99y on AIX
Sometimes a process with (CPU) TIME reported as '-' appears in
the 'ps' output on AIX 7. It's correctly not matched by the
filter including processes with TIME between 0 seconds and 99
years because if we cannot tell, not including the process seems
to be the safest thing we can do.

Our test checks that there is no such process that doesn't fall
into the 0s-99y boundaries so we need to make it less strict on
AIX.

Ticket: ENT-4487
Changelog: None