prefer routable IPv6 addresses over link-local
Currently the IPv6 address provided for an interface in the $(sys.inet6)
data is arbitrary if more than one address is bound to that interface.
Given that IPv6 requires all interfaces have a link-local address, this
is a common scenario.  Add a mechanism to GetProcFileInfo() to choose
between the previous and new entries if more than one is found for an
extracted key, and use this mechanism to rank link-local addresses below
others.

Also, the post-processing functions passed to GetProcFileInfo() have a
strange return value.  This is seemingly due to a misunderstanding with
how function pointer types relate to function definitions.  Clean this
up.