Use the shortest/nicest tag pointing at recent_rev in determine_version.py
Quite often there are multiple tags pointing at the most recent
tagged revision. If that's the case, we want to use the
shortest/nicest tag to determine whether to increment the minor
or patch version number and to what we should increment it.

'git tag --points-at REVISION' lists all tags pointing at the
given revision and the list is sorted just like we need it. For
example:

  3.12.2
  3.12.2-2
  3.12.2-build2
  3.12.2-build3
  3.12.2-build4
  3.12.2-build5