Use python3 for determine_version.py
It's compatible with Python 2.7 and Python 3, but it's better
(required on RHEL 8) to be explicit. And now that we have to
choose, Python 3 is a better option.

determine_version.py is run as part of ./autogen.sh which is a
build stage we run on modern OS in Jenkins and so should be
people running it locally.

Ticket: CFE-3039
Changelog: None