LGTM: Missing null check query now looks for asserts
Looks to me like the LGTM build runs without debug enabled,
so the asserts are not expanded to comparisons. This commit
adds the additional check for a macro invocation where macro
name is `assert` and the "argument", the unexpanded string
inside parenthesis, matches the expected string. A little hacky,
but it gets rid of 600 false positives.

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>