Fixed multiple measurements tracking growth of same file
Use handle+filename to track running file position.
Previously the key was the filename, this is problematic in the case of multiple
measurements from the same growing file. handle+filename is more specific
handling multiple promises for the same file as well as the file for a given
handle changing. Promise handle would be sufficient to cover the case of
multiple measurements from the same file but it has an edge case in the event
the filename changed but the handle remained the same.

Ticket: ENT-4814
Changelog: Title
(cherry picked from commit 8c5bc869d4c4786138cf1fb33f9665b28735e753)