Fix floating point number parsing in eval()
For example, 0.75 can also be written as .75 and should thus be
parsed into the same value.

The relevant change in this commit is the small one in the
math.peg file (the grammar), math.pc is generated from it with
the 'peg' tool from the 3rdparty/peg directory.

Ticket: CFE-2762
Changelog: Fixed .xy floating point numbers parsing in eval()