Added unit test and fixed issue with CheckParseVariableName
I wanted to only add a unit test, but discovered that the
result depends on the contents of an uninitialized buffer
for one edge case, so I felt like I had to fix it.

I want this unit test so I can change(fix) the implementation
with some confidence.

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>