cf-check: dump command now dumps DB contents to JSON5
Example:

```
$ cf-check dump /var/cfengine/state/cf_lastseen.lmdb
{
        "a192.168.100.10\0": "MD5=efbb78408659e9e99188218a8518b7ec\0",
        "kMD5=efbb78408659e9e99188218a8518b7ec\0": "192.168.100.10\0",
        "qoMD5=efbb78408659e9e99188218a8518b7ec\0": "\0\xf6h]\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0",
}
```

`dump` was previously an alias for `lmdump`.

`lmdump` command, as well as symlinking `cf-check` as `lmdump`,
is still supported, for compatibility.

New functionality will be added to `dump` command.

Some code should be moved to JSON library, but I need to make
bigger changes there, so I think it's okay to merge this first.

Changelog: Title
Ticket: CFE-3126
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>