Tell LMDB to call our function when it hits a failing assert
A failing assert in LMDB for us means that the LMDB file is
corrupted and actions need to be taken to mitigate the
situation. Fortunately, we can tell LMDB to call our function
whenever it hits an assert.

For now, this function only logs a message identifying which DB
got corrupted. In the future, we want this handler to do much
more, ideally transparently handle the situation.

Ticket: CFE-2469
Changelog: Report corrupted local database with a critical log message