Added CheckClassExpression and CheckVarClassExpression
IsClassDefined and IsVarClassDefined use boolean return type
and do not distinguish between failure and false.

As an example, this can be problematic/ambiguous:

```
if (!IsVarClassDefined())
```

(Is the class not defined or was an error encountered?)

These new functions can be used in situations like this.
EvalClassExpression was already taken, so I made up different
names (Check). This needs further cleanup.

Changelog: None
Ticket: CFE-2689
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>