Keys in the lock database, cf_lock.lmdb, are now human-readable
LMDB keys are limited to 511 bytes, by default. To get around
this, the keys were changed to be hashes some time ago.
This change effectively reverts that commit, making the keys
normal strings by default, and overwriting the last part with a
hash, if necessary (if the key would exceed 511 bytes).

Changelog: Title
Ticket: CFE-2596
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>