Try to repair corrupted LMDB files in reaction to corruption detection
'cf-agent' and some of our daemons check if LMDB files are OK
when starting. However, this check is not good enough to detect
all the problems (it's just trying to read the DB contents) and
corruptions can also happen while the process is running.

We should try to repair the corrupted LMDB files when the
corruption is detected and reported by LMDB itself (on an
assertion failure in LMDB).

Ticket: CFE-3127
Changelog: Corrupted LMDB files are now automatically repaired