Fix handling of the 'corrupt' argument of diagnose_files()
The `Seq **corrupt` argument of the diagnose_files() function is
an out parameter, a place to store the sequence of detected
corrupted files. It can be NULL if the caller only wants to get
the number of such corrupted files, but otherwise it doesn't have
to be initialized. In fact, it should not be initialized in most
cases to not confuse the "used but not initialized" compiler
check.

Instead of `assert(corrupt == NULL | *corrupt == NULL)`, let's
just decribe the meaning of parameters and return value in a
doxygen comment. And make sure a new sequence is created for the
out value.

Also mark the first parameter (`Seq *files`) as `const`, because
the sequence is not modified by the function.

(cherry picked from commit cf25b56157ddaf6e8f47a3bd7270a12385f4174f)