Use replicate_lmdb_file() in 'cf-check repair'
'cf-check repair' should try to preserve the data that can be
preserved. So if an LMDB file is corrupted, it should try to read
as much as it can from it and write it into a new file and then
replace the original file with the new copy. Hence
`replicate_lmdb_file()` and then `rename(orig, new_copy)`.

Also, add a new '--force' option to 'cf-check repair' to repair
all the files no matter if they are reported as corrupted or not.

Ticket: CFE-3127
Changelog: 'cf-check repair' now preserves readable data in corrupted LMDB files