two 'Makefile.am' under 'tests/' need adjustment for "make check"
Without this, "make check" fails to build some programs, claiming
missing dependencies.  With it, the builds work.

Given the nature of this change, however, I suspect that this particular fix
may be incorrect. If those source->object recompilations (as distinct
from using already-built objects) really are necessary, then a variation of
the original lines will need to be reinstated.

Does this tie in with the recent splitting out of the "libntech" submodule?
And some sort of "PATH" (or similar) detail mismatch when that is built
as a subtree of "cfengine"?