Network protocol version 2 is now called "tls"
"tls" or "2" can be used in places where you specify network
protocol. Log messages were altered, to show "tls" instead of
"latest". "latest" can still be used, but it will be an
alias for whatever is the latest version. This change was made
to pave the way for new protocol versions and features.

Changelog: Commit
Ticket: ENT-4406
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>