Added NULL check assertions for ConnectionInfo pointers
It seems arbitrary that these functions accept NULL pointers.
As an example, what should `ServerTLSPeek(NULL)` do?
It seems like the only cases where this would happen are when
(connection) errors occur. It is better to assert in these
cases and fix the error handling if it breaks somewhere.
Kept the release build handling as is, in fear of breaking
things.

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>