revert to original commit: the early 'assert'
In this flurry of commits, the original had the 'assert' early in the function.
But the test phase of some subsequent work (on a separate branch) raised
the suspicion that there may have been a problem.
So for safety I moved the 'assert' lower in the function, to be just prior
to first use of the pointer (the second commit).

I then realised that my suspicion may have been groundless.
So I moved the 'assert' back to its first location (in this third commit).

It continues to pass the tests properly.

So the net effect is the single, original commit.