Introduced protocol version 3 - cookie
This protocol is identical to version 2 ("tls"),
except it allows the enterprise reporting hub to send
the COOKIE command to enterprise hosts. This command is used for
detecting hosts using duplicate identities.

For community installations, it should not make a difference,
policy servers will not send this command. The only visible
difference is the new version number (in logs and policy).

Changelog: Commit
Ticket: ENT-4406
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>