Changed unless constraint to be more consistent with if
For any situation where `if` would NOT skip a promise, `unless`
will cause the promise to be skipped. When there are
unresolved variables / function calls, `if` will skip, `unless`
will NOT skip.

Both of these keywords make a determination on whether to SKIP
the promise or not. If used together skipping effectively takes
precedence. In other words, a promise will be evalutated, if
nothing indicates that it should be skipped.

Tests and log messages were also updated.

Changelog: Commit
Ticket: CFE-3160
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>