Fixed failure to build when source and build locations differ
The CFEngine code used to get this right, but a couple of issues have crept in
at some point.

Although most people build in the same directory-tree location as the source,
GNU autotools offers the ability for these locations to differ.
This is particularly useful when wanting to develop and test parallel builds
from a shared source location, for instance using different platforms and/or
different "--with-foo" options.

A typical use would be the following parallel builds:

    mkdir -p build/platform_1; cd build/platform_1; ../../autogen.sh
    mkdir -p build/platform_2; cd build/platform_2; ../../autogen.sh
    mkdir -p build/platform_3; cd build/platform_3; ../../autogen.sh

Subsequent development in the source can then be immediately tested
on those three platforms, in parallel, and with the previous cached 'make'
results within each still valid.  This is much, much faster!

This fix restores this capability.

There is one residual exception, outside CFEngine, but still easily fixable.
In the git submodule "libntech", file "libntech/tests/unit/csv_parser_test.c"
needs to use the "TESTDATADIR" macro in two places in a way very similar
to the nearby "json_test.c".  (I'm not quite sure how to submit that
particular fix as a pull-request, but can easily provide a 'diff' on request.)