Add a recommended minimum TLS version and default to it
Security vulnerabilities were found in TLS 1.0 [1]. None of them
applies to the mutually authenticated peer-to-peer TLS
communication in CFEngine, but the version 1.0 of the protocol is
now considered deprecated.

TLS 1.1 is supported by CFEngine 3.7.0 and newer so it is safe to
default the minimum required TLS version to 1.1 while making it
possible to change the minimum required version to 1.0 in policy.

[1] https://www.acunetix.com/blog/articles/tls-vulnerabilities-attacks-final-part/

Ticket: ENT-4616
Changelog: Default minimum allowed TLS version is now 1.1