CFE-3046: Only add unique IPv4 interfaces to sys.interfaces
Previously, when an interface had multiple IPv4 addresses set, the parsing of
these addresses would add the interface to the sys.interfaces and sys.hardware
variables multiple times, resulting in the interface showing up as a duplicate
in the list.

This fix makes sure that an interface is only added once for every interface
type, regardless of how many IPv4 addresses the interface contains.

Changelog: Fixed an issue causing duplicate entries in sys.interfaces, and
sys.hardware.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>