Mitigate race condition in the promiser_match_is_correct.cf test
We spawn an agent process in a shell and then immediately check
that it is running. If it takes a bit to start (easy to happen on
a heavy-loaded machine running tests) the process is not
found. Let's give it a second to start before we try extracting
information about it.

(cherry picked from commit a623c13507d6d0620d11104f3c86dffedfcbdf66)