Aligned style between an example that renders properly, and one that does not
The function-return-types.cf example is rendered in the Functions section of the
reference manual.

https://docs.cfengine.com/docs/master/reference-functions.html

It renders the prep section, the policy, and the output sections in blocks.

files_depth_search_include_basedir.cf is rendered in the include_basedir
section.

https://docs.cfengine.com/docs/master/reference-promise-types-files.html#include_basedir

It renders the prep section poorly.

Ticket: CFE-3217
Changelog: None