Fixed small memory leak in policy evaluation
This memory leak appears when the LHS of a promise
constraint is `data` or `template_data`, and the RHS
is a string literal. If the string literal is not
valid JSON, perhaps because of unexpanded variables,
a `parsejson` function call is inserted instead. The name
of the function call (for example `parsejson`) is leaked
in this case.

Not picked up by our CI, since valgrind test only runs
community parts of masterfiles.

Changelog: Title
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>