Moved error reading file in countlinesmatching() from verbose to error
countlinesmatching() did not error when the file being counted does not exist or
was unable to be accessed. The result of 0 matching lines is correct, but there
is also an expectation of an error message to indicate the file is missing.

Added test for countlinesmatching() erroring on absent file.

This test simply runs an existing test and inspects the agent run output for an
error.

Ticket: CFE-3234
Changelog: Title