Find the matching parenthesis in variable reference
If a variable reference contains parentheses, we need to make
sure we find the matching parenthesis for the opening one, not
the first opposite parenthesis.

For example '$(my_array[key(1)])' needs to match the second
closing parenthesis not the first one.

Ticket: CFE-3242
Changelog: Variable references with nested parentheses no longer
cause errors