cf-check: Added minimal validation for cf_changes.lmdb
This database has non-string keys and values, and
has not been known to cause a lot of issues, so I
added a separate mode for very simplistic validation;
only read the binary data.

No changelog entry since this is part of --validate,
another commit with changelog entry and it was not
released yet.

Changelog: None
Ticket: CFE-2988
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>