cf-check: Fixed child process exit codes
On at least some systems (my Ubuntu 18 VM) exit codes can only
be in the range 0-255. Thus exiting with -1 or -32000 is bad.
The most sane thing to do is to exit with our custom cf-check
error codes, which range from 0 to approximately 60, and then
extended by system errno codes which are typically 0 - ~100.

Note that signals are handled separately, exit code is only
used if the process was not terminated by signals (SIGBUS,
SIGSEGV, SIGABRT, ...).

Changelog: None
Ticket: CFE-2988
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>