cf-agent: files: report purged dirs and files as repaired
Given a recursive file promise with purging enabled, the promise will
not be propagated as repaired if the target directory contains some
directory hierarchy that is removed by the promise. While the function
`PurgeLocalFiles` correctly keeps track of any failed attempts to remove
files from the target directory, it does not report the repaired promise
in case it removes files or directories from the target directory
successfully.

Fix the issue by properly updating the promise to PROMISE_RESULT_CHANGE
if puring either a directory or file successfully. Add an acceptance
test to verify behaviour.

Changelog: Title
Ticket: CFE-3260
Signed-off-by: Patrick Steinhardt <ps@pks.im>