CFE-2898: New data validation policy functions validdata and validjson
These functions take data as a string, and return true if the data was
successfully parsed, and false if an error was encountered. They will
print whatever the error was as LOG_LEVEL_INFO.

Also added acceptance tests. The functionality is expected to be
expanded to support CSV, YAML and ENVFILEs as well.

Ticket: CFE-2898
Changelog: Title