cf-check: Symlinked LMDB databases are now preserved in repair
Performs diagnosis and repair on symlink target instead of symlink.
Repaired files / copies are placed alongside symlink target.
In some cases, the symlink target is deleted to repair a corrupt
database, and the symlink is left as a broken symlink. This is
handled gracefully by the agent, it will be recreated. Broken
symlinks are now detected as an acceptable condition in diagnose,
it won't try to repair them or delete them.

Changelog: Commit
Ticket: ENT-5162
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>