Do proper encryption using RSA+AES in cf-keycrypt
Encrypting just with the RSA key is slow and also potentially
insecure because the security of the encryption with RSA degrades
with the data being bigger than the RSA key itself.

The standard way of encryption using RSA is to generate a AES
key, encrypt the key with the RSA key and then encrypt the actual
payload with the generated AES key using symmetric encryption. Of
course, the encrypted AES key then has to be stored next to the
encrypted payload for decryption. The same applies to a randomly
generated initialization vector (IV) needed by the AES cipher.

Ticket: CFE-2613
Changelog: none
(cherry picked from commit f27b70a3b65ef157a4dcbe200459090e7503af31)