Do not log an error in case of no match for a measurements promise
If a measurements promise regex doesn't match any line of a given
file, it should not produce an error logged to syslog. It's just
noise and the default value 0.0 of the measurement indicates that
nothing was measured. Running with '--verbose' then shows the log
message and helps with debugging.

Ticket: ENT-5171
Changelog: Measurements promises with no match no longer produce errors