Always set $(sys.libdir) and $(sys.local_libdir)
These no longer depend on version so they should be set in all
cases not just when the CFEngine version is known and parsed.

The code should not be in DiscoverVersion(), but that's out of
scope of the changes done here.

(cherry picked from commit a8aa6274c98fa2b3de99c666a1b96edb50cc9a57)