cf-remote: Fixed issue with installing MSI on windows 2012 server
Sleeping inside the ssh command allows the installer to finish,
on 2012 it would be killed when the ssh connection was closed.

A more robust solution would be to change our MSI to not
exit immediately (if possible), or to transfer a script which
monitors and waits for installation to finish, but this works
for now.

Tested on our windows 2012 and 2016 AMIs. One is powershell,
the other is cmd.

Ticket: None
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>