Add a new secret Rval value
Standard Rval value which represents all the secrets. We need
some *static* value that we can return instead of real Rval
values in cases where the real Rval value is tagged as
secret. This way we don't have to be concerned about memory
leaks when returning the fake "**********" Rval instead of the
real value.

Ticket: CFE-3297
Changelog: None