Renamed promise_type to promise_guard in parser
promise_type is going to be very ambiguous when you can
define custom promise types (Are we talking about
the promise types in a bundle or the tokens when defining a new
promise type?). I opted for promise_guard because it looks
similar to a class guard, even though it doesn't "guard" anything.

Some alternatives:
* `set_promise_type`
* `promise_type_start`
* `set_current_promise_type`
* `promise_type_begin`

All of those are more technically correct, but I think they are
somewhat ambiguous.

Changelog: None
Ticket: CFE-3273
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>