Renamed PromiseType struct to BundleSection
"PromiseType" is ambiguous, is it just the part before the
colon? Including the colon? No, it is actually the promise type,
and the colon, and all of the following promises.

So I decided to name this a bundle section, or section for short.
A section is a promise guard (promise type + colon), followed by
0 or more promises, which may have class guards.

A lot of "Find & Replace" in editor, and then manually reviewing
and editing all of the touched places. I also updated function
names and local variable names to reflect this. I did so in the
same commit, because the change in the type name showed me
the places that should be updated.

I intentionally left some struct member names, like
`parent_promise_type`, and `promise_types` because they will be
easy to replace in a separate commit, making it easier to review
and hopefully less error-prone.

Changelog: None
Ticket: CFE-3273
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>